backend/python/requirements.txt
# Python Backend Services - Unified Requirements
# Compatible with Python 3.8+

# ==================================================
# FastAPI and Server
# ==================================================
fastapi>=0.70.0,<1.0.0
uvicorn>=0.15.0,<1.0.0
python-multipart>=0.0.5

# ==================================================
# Data Processing (SmartBI)
# ==================================================
pandas>=1.3.0,<2.0.0
openpyxl>=3.0.0,<4.0.0
numpy>=1.20.0,<2.0.0
xlrd>=2.0.0,<3.0.0

# ==================================================
# Statistical Analysis and Forecasting (SmartBI)
# ==================================================
statsmodels>=0.13.0,<1.0.0
scipy>=1.7.0,<2.0.0

# ==================================================
# LLM Integration (SmartBI + Efficiency Recognition)
# ==================================================
httpx>=0.20.0,<1.0.0
aiohttp>=3.8.0,<4.0.0
openai>=1.0.0  # For OpenAI-compatible API (DashScope, Qwen)

# ==================================================
# Data Validation
# ==================================================
pydantic>=1.9.0,<2.0.0
pydantic-settings>=2.0.0 ; python_version >= "3.9"

# ==================================================
# Database (PostgreSQL)
# ==================================================
sqlalchemy>=1.4.0,<2.0.0
psycopg2-binary>=2.9.0,<3.0.0

# ==================================================
# NVR/Video Processing (Efficiency Recognition)
# ==================================================
opencv-python>=4.8.0,<5.0.0

# ==================================================
# Task Scheduling (Multi-stream Sampler)
# ==================================================
apscheduler>=3.10.0,<4.0.0

# ==================================================
# Utilities
# ==================================================
python-dateutil>=2.8.0,<3.0.0
typing-extensions>=3.7.4
python-dotenv>=0.19.0
